feat: S3, AWS 관련 config 수정
