+++ mpl-domain/src/main/java/com/codeit/sb02mplteam2/config/S3Config.java
@@ -1,45 +1,22 @@
-package com.codeit.sb02mplteam2.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.presigner.S3Presigner;
-
-@Configuration
-public class S3Config {
-
-  @Value("${mpl.aws.accessKeyId}")
-  private String accessKeyId;
-
-  @Value("${mpl.aws.secretKey}")
-  private String secretKey;
-
-  @Value("${mpl.aws.region}")
-  private String region;
-
-  @Bean
-  public S3Client s3Client() {
-    AwsBasicCredentials awsCredentials = AwsBasicCredentials.create(accessKeyId, secretKey);
-
-    return S3Client.builder()
-        .region(Region.of(region))
-        .credentialsProvider(StaticCredentialsProvider.create(awsCredentials))
-        .build();
-  }
-
-  @Bean
-  public S3Presigner s3Presigner() {
-    AwsBasicCredentials awsCredentials = AwsBasicCredentials.create(accessKeyId, secretKey);
-
-    return S3Presigner.builder()
-        .region(Region.of(region))
-        .credentialsProvider(StaticCredentialsProvider.create(awsCredentials))
-        .build();
-  }
-
-}
-
+//package com.codeit.sb02mplteam2.config;
+//
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+//import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+//import software.amazon.awssdk.regions.Region;
+//import software.amazon.awssdk.services.s3.S3Client;
+//import software.amazon.awssdk.services.s3.presigner.S3Presigner;
+//
+//@Configuration
+//public class S3Config {
+//
+//  @Bean
+//  public S3Client s3Client() {
+//    return S3Client.builder()
+//        .region(Region.AP_NORTHEAST_2)
+//        .build();
+//  }
+//
+//}
